running screenshot tests on ios-safari unit tests (flutter/engine#20963)
* running screenshot tests on ios-safari unit tests

* fixing the golden_smoke tests. changes to the documentation

* addressing reviewer comments

* cropping footer from the simulator screenshot. addressing some reviewer comments

* use .dart_tools for recording the screenshots

* fix the usage of the method

* adding TODO's for missing documentation and not supported windows tests

* addressing comments

* changing to incremental counter for file names

* add comment to the counter

* fix anaylze issues

* using takescreenshot method from the iosSimulator.

* address reviewer comments

* fix the scaling issue. disable eronous test

* change the smoke file for top gap 282

* change the variable name for scale factor

+++ engine/src/flutter/lib/web_ui/dev/utils.dart
@@ -28,8 +28,7 @@ class FilePath {
 
   @override
   bool operator ==(Object other) {
-    return other is FilePath
-        && other._absolutePath == _absolutePath;
+    return other is FilePath && other._absolutePath == _absolutePath;
   }
 
   @override
@@ -216,7 +215,9 @@ void cleanup() async {
   // Delete temporary directories.
   if (temporaryDirectories.length > 0) {
     for (io.Directory directory in temporaryDirectories) {
-      directory.deleteSync(recursive: true);
+      if (!directory.existsSync()) {
+        directory.deleteSync(recursive: true);
+      }
     }
   }
